:truck: move getOptions
